Created Harmony Patch for HeroCreator
This is a simple patch that modifies the result of the base game "DeliverOffSpring(...)" method so that the clan of the child will depend on the mother's clan, allowing bastard children and preventing children teleporting across the map once being born. At this time, it does not have MCM integration.