Compile Error Fixes
I was finally able to get Visual Studio up so I can compile .dll files to test my edits (after a million dependencies were added to the solution)

This was able to compile and run in the latest version of Bannerlord and I tested the functionality of the console commands. It is not guaranteed to work each time you run it due to how the ImpregnationChance() method works. I set Pregnancy to 100% chance for testing purposes