Release 1039 (#83)
* Updated MCM

* Added a TryCatch

* Updated for V102

* Fixed Specific Bug

* Fixed Leveling Issue

* Fixed to StripSettings

* Fixed Minor Bug on Sell

* Added "captivity.ImpregnateBy" Console Command (#76)

* Added "captivity.ImpregnateBy" Console Command

This console command is a modification of "Impregnant" which takes a second value in order to search for a hero to serve as the "father" for the target hero being impregnated. There is no gender check in this method as the "ImpregnationChance(...)" method contains this check.

If a female hero is entered as the "fromName" input then the method will complete and call the "ImpregnationChance" method however with it's current behavior it will default the father to be a randomly created hero. In order to allow for female-female impregnation then that method must be altered.

Co-authored-by: DMRTYR <120136261+DMRTYR@users.noreply.github.com>

* Added a few Bug Fixes

* Fixed flagValid

* NeutralFaction

* New Item Capability

* Fixed Minor Bugs

* Formatting

* Fixes a Number of Issues

* Version v1.1.0

* MCM + Female-Female Integration (#78)

* PregnancyToggleFemalexFemale

This inserts the "PregnancyToggleFemalexFemale" boolean into the custom settings

* Updated CESettings to include PregnancyToggleFemalexFemale

Adds the boolean PregnancyToggleFemalexFemale to the MCM to control behavior

* Implemented Gendercheck Toggle

Inserts a check at the point that filters out females as "senders"/"captors" that checks if they are female and if female-female impregnation is allowed

* Fix incorrect parameter

The "forcepreg" parameter in ImpregnationChance currently does not allow for a selected father and defaults to a random one. By switching it to false, the pregnancy is no longer guaranteed but you can control the father

* Created Harmony Patch for HeroCreator

This is a simple patch that modifies the result of the base game "DeliverOffSpring(...)" method so that the clan of the child will depend on the mother's clan, allowing bastard children and preventing children teleporting across the map once being born. At this time, it does not have MCM integration.

* Compile Error Fixes

I was finally able to get Visual Studio up so I can compile .dll files to test my edits (after a million dependencies were added to the solution)

This was able to compile and run in the latest version of Bannerlord and I tested the functionality of the console commands. It is not guaranteed to work each time you run it due to how the ImpregnationChance() method works. I set Pregnancy to 100% chance for testing purposes

* Fixed of Postfix

* Fixed #79

* Fixed Minor Patch

---------

Co-authored-by: DMRTYR <120136261+DMRTYR@users.noreply.github.com>